Rename new processor files, functions and types
`makeNewProcessor` becomes `makeProcessor`, `NewProcessor` becomes
simply `Processor`, etc.
